# api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080/weather","description":"Generated server url"}],"paths":{"/api/warnings/receive":{"post":{"tags":["warning-controller"],"operationId":"receiveWarning","parameters":[{"name":"warningType","in":"query","required":true,"schema":{"type":"string"}},{"name":"warningLevel","in":"query","required":true,"schema":{"type":"string"}},{"name":"content","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/warnings/publish":{"post":{"tags":["admin-controller"],"operationId":"publishComprehensiveWarning","parameters":[{"name":"warningId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/warnings/create":{"get":{"tags":["admin-controller"],"operationId":"createComprehensiveWarningPage","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}},"post":{"tags":["admin-controller"],"operationId":"createComprehensiveWarning","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/WarningPublishDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/login":{"post":{"tags":["admin-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/warnings":{"get":{"tags":["weather-controller"],"operationId":"weatherWarnings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/warnings/type":{"get":{"tags":["weather-controller"],"operationId":"warningsByType","parameters":[{"name":"warningType","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/warnings/level":{"get":{"tags":["weather-controller"],"operationId":"warningsByLevel","parameters":[{"name":"warningLevel","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/warning/detail":{"get":{"tags":["weather-controller"],"operationId":"warningDetail","parameters":[{"name":"id","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/realtime":{"get":{"tags":["weather-controller"],"operationId":"realtimeWeather","parameters":[{"name":"location","in":"query","required":false,"schema":{"type":"string","default":"SANYA"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/info":{"get":{"tags":["home-controller"],"operationId":"weatherInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/history":{"get":{"tags":["weather-controller"],"operationId":"weatherHistory","parameters":[{"name":"location","in":"query","required":false,"schema":{"type":"string","default":"SANYA"}},{"name":"date","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/forecast":{"get":{"tags":["weather-controller"],"operationId":"weatherForecast","parameters":[{"name":"location","in":"query","required":false,"schema":{"type":"string","default":"SANYA"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather/compare":{"get":{"tags":["weather-controller"],"operationId":"weatherCompare","parameters":[{"name":"location1","in":"query","required":false,"schema":{"type":"string"}},{"name":"location2","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/weather-map":{"get":{"tags":["home-controller"],"operationId":"weatherMap","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/warnings/summary":{"get":{"tags":["warning-controller"],"operationId":"warningSummary","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/warnings/subscription":{"get":{"tags":["warning-controller"],"operationId":"warningSubscription","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/warnings/info":{"get":{"tags":["home-controller"],"operationId":"warningsInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/warnings/comprehensive":{"get":{"tags":["warning-controller"],"operationId":"comprehensiveWarnings","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/warnings/comprehensive/{id}":{"get":{"tags":["warning-controller"],"operationId":"comprehensiveWarningDetail","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/warnings/analysis":{"get":{"tags":["warning-controller"],"operationId":"warningAnalysis","parameters":[{"name":"warningId","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/traffic/trains":{"get":{"tags":["traffic-controller"],"operationId":"trainStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/traffic/search":{"get":{"tags":["traffic-controller"],"operationId":"trafficSearch","parameters":[{"name":"type","in":"query","required":false,"schema":{"type":"string"}},{"name":"city","in":"query","required":false,"schema":{"type":"string"}},{"name":"startDate","in":"query","required":false,"schema":{"type":"string"}},{"name":"endDate","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/traffic/realtime":{"get":{"tags":["traffic-controller"],"operationId":"realtimeTraffic","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/traffic/info":{"get":{"tags":["home-controller"],"operationId":"trafficInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/traffic/impact":{"get":{"tags":["traffic-controller"],"operationId":"trafficImpact","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/traffic/flights":{"get":{"tags":["traffic-controller"],"operationId":"flightStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/traffic/detail":{"get":{"tags":["traffic-controller"],"operationId":"trafficDetail","parameters":[{"name":"type","in":"query","required":true,"schema":{"type":"string"}},{"name":"number","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/index":{"get":{"tags":["home-controller"],"operationId":"index","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/help":{"get":{"tags":["home-controller"],"operationId":"help","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/status":{"get":{"tags":["attraction-controller"],"operationId":"attractionsStatus","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/search":{"get":{"tags":["attraction-controller"],"operationId":"attractionSearch","parameters":[{"name":"name","in":"query","required":false,"schema":{"type":"string"}},{"name":"status","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/recommendations":{"get":{"tags":["attraction-controller"],"operationId":"weatherRecommendations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/open-today":{"get":{"tags":["attraction-controller"],"operationId":"openToday","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/map":{"get":{"tags":["attraction-controller"],"operationId":"attractionsMap","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/info":{"get":{"tags":["home-controller"],"operationId":"attractionsInfo","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/detail":{"get":{"tags":["attraction-controller"],"operationId":"attractionDetail","parameters":[{"name":"name","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/attractions/by-location":{"get":{"tags":["attraction-controller"],"operationId":"attractionsByLocation","parameters":[{"name":"locationCode","in":"query","required":false,"schema":{"type":"string","default":"SANYA"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/weather-data":{"get":{"tags":["admin-controller"],"operationId":"weatherDataManagement","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/warnings":{"get":{"tags":["admin-controller"],"operationId":"warningManagement","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/users":{"get":{"tags":["admin-controller"],"operationId":"userManagement","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/traffic":{"get":{"tags":["admin-controller"],"operationId":"trafficManagement","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/logs":{"get":{"tags":["admin-controller"],"operationId":"systemLogs","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/logout":{"get":{"tags":["admin-controller"],"operationId":"logout","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/dashboard":{"get":{"tags":["admin-controller"],"operationId":"dashboard","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/admin/attractions":{"get":{"tags":["admin-controller"],"operationId":"attractionManagement","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/about":{"get":{"tags":["home-controller"],"operationId":"about","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}},"components":{"schemas":{"WarningPublishDTO":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"},"warningLevel":{"type":"string"},"affectedLocations":{"type":"array","items":{"type":"string"}},"weatherWarningIds":{"type":"array","items":{"type":"integer","format":"int64"}},"trafficEffects":{"type":"string"},"attractionEffects":{"type":"string"}}},"UserLoginDTO":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}}}}}